Make energy double everywhere

+++ src/main/java/org/jetbrains/car/ElectricCar.java
@@ -4,7 +4,7 @@ import org.jetbrains.location.Location;
 import org.jetbrains.util.Constants;
 
 public class ElectricCar extends Car {
-    public ElectricCar(Location location, int energyUsageRate) {
+    public ElectricCar(Location location, double energyUsageRate) {
         super(location, energyUsageRate);
         energyThreshold = Constants.ELECTRIC_CAR_ENERGY_THRESHOLD;
     }
